# ScSoMe.Common/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ScSoMe.API",
    "version": "1.0"
  },
  "paths": {
...
    "/Message/CheckChatExist": {
      "get": {
        "tags": [
          "Message"
        ],
        "operationId": "CheckChatExist",
        "parameters": [
          {
            "name": "senderId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "receiverId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "groupName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "displayName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DcHistory"
                }
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/Message/SaveMessage": {
      "post": {
        "tags": [
          "Message"
        ],
        "operationId": "SaveMessage",
        "parameters": [
          {
            "name": "groupName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "text",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "senderId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "senderName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "200": {
            "description": "Success"
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/Message/GetChatMessageHistory": {
      "get": {
        "tags": [
          "Message"
        ],
        "operationId": "GetChatMessageHistory",
        "parameters": [
          {
            "name": "groupName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageHistory"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/Message/GetLastChatMessage": {
      "get": {
        "tags": [
          "Message"
        ],
        "operationId": "GetLastChatMessage",
        "parameters": [
          {
            "name": "groupName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageHistory"
                }
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/Message/GetAllChatsByUser": {
      "get": {
        "tags": [
          "Message"
        ],
        "operationId": "GetAllChatsByUser",
        "parameters": [
          {
            "name": "currentUserId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DcHistory"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/Message/MarkMessagesAsRead": {
      "put": {
        "tags": [
          "Message"
        ],
        "operationId": "MarkMessagesAsRead",
        "parameters": [
          {
            "name": "groupName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "currentUserId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/Message/GetUnreadMessageCount": {
      "get": {
        "tags": [
          "Message"
        ],
        "operationId": "GetUnreadMessageCount",
        "parameters": [
          {
            "name": "groupName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "currentUserId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/Message/GetTotalUnreadMessageCount": {
      "get": {
        "tags": [
          "Message"
        ],
        "operationId": "GetTotalUnreadMessageCount",
        "parameters": [
          {
            "name": "currentUserId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/Message/DeleteChat": {
      "delete": {
        "tags": [
          "Message"
        ],
        "operationId": "DeleteChat",
        "parameters": [
          {
            "name": "groupName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/Message/DeleteSingleMessage": {
      "delete": {
        "tags": [
          "Message"
        ],
        "operationId": "DeleteSingleMessage",
        "parameters": [
          {
            "name": "groupName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "message",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "senderId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/Message/ChangeChatDisplayName": {
      "put": {
        "tags": [
          "Message"
        ],
        "operationId": "ChangeChatDisplayName",
        "parameters": [
          {
            "name": "groupName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "newDisplayName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/Message/CheckIfDeleted": {
      "get": {
        "tags": [
          "Message"
        ],
        "operationId": "CheckIfDeleted",
        "parameters": [
          {
            "name": "groupName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DcHistory": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "groupName": {
            "type": "string",
            "nullable": true
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "newDisplayName": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "MessageHistory": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "chatId": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "senderId": {
            "type": "integer",
            "format": "int32"
          },
          "senderName": {
            "type": "string",
            "nullable": true
          },
          "isRead": {
            "type": "string",
            "nullable": true
          },
          "sentDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}